requirements.txt
APScheduler==3.9.1.post1
Flask==2.2.3
opencv_python==4.7.0.72
PyDrive==1.3.1
python-dotenv==1.0.0
Requests==2.28.2
tweepy==4.13.0
